Defined custom asymmetric quantization operator signatures
